+++ server/src/test/scala/stasis/test/specs/unit/server/api/routes/StagingSpec.scala
@@ -112,7 +112,7 @@ class StagingSpec extends AsyncUnitSpec with ScalatestRouteTest {
 
   private val testManifest = Manifest(
     crate = Crate.generateId(),
-    size = testContent.size,
+    size = testContent.size.toLong,
     copies = 4,
     source = Node.generateId(),
     origin = Node.generateId()
